backend/requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
sqlalchemy>=2.0.23
psycopg2-binary>=2.9.9
pydantic>=2.5.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dotenv>=1.0.0
websockets>=12.0
python-socketio>=5.10.0
email-validator>=2.1.0
httpx>=0.27.0
google-generativeai>=0.3.0
pywebpush>=1.18.0
numpy>=1.24.0
scikit-learn>=1.3.0
filterpy>=1.4.5
pyjwt>=2.8.0
requests>=2.31.0